# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreatePaymentRecipientRequest:
    properties:
      accountNumber:
        type: string
      name:
        type: string
    required:
    - accountNumber
    - name
    type: object
  handlers.CreatePaymentRequest:
    properties:
      amount:
        type: number
      fromAccount:
        type: string
      paymentCode:
        type: string
      purpose:
        type: string
      recipientAccount:
        type: string
      recipientName:
        type: string
      referenceNumber:
        type: string
    required:
    - amount
    - fromAccount
    - recipientAccount
    - recipientName
    type: object
  handlers.CreateTransferRequest:
    properties:
      amount:
        type: number
      fromAccount:
        type: string
      toAccount:
        type: string
    required:
    - amount
    - fromAccount
    - toAccount
    type: object
...
  handlers.UpdatePaymentRecipientRequest:
    properties:
      accountNumber:
        type: string
      name:
        type: string
    required:
    - accountNumber
    - name
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: REST API gateway for EXBanka microservices.
  title: EXBanka API Gateway
  version: "1.0"
paths:
...
  /api/payments:
    get:
      description: Returns all payments made from the authenticated client's accounts,
        with optional filters.
      parameters:
      - description: From date (RFC3339)
        in: query
        name: date_from
        type: string
      - description: To date (RFC3339)
        in: query
        name: date_to
        type: string
      - description: Min amount
        in: query
        name: amount_min
        type: number
      - description: Max amount
        in: query
        name: amount_max
        type: number
      - description: Payment status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List client payments
      tags:
      - payments
  /api/payments/{paymentId}:
    get:
      description: Returns details of a single payment owned by the authenticated
        client.
      parameters:
      - description: Payment ID
        in: path
        name: paymentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get payment details
      tags:
      - payments
  /api/payments/create:
    post:
      consumes:
      - application/json
      description: Initiates a payment from client's account to a recipient account.
      parameters:
      - description: Payment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new payment
      tags:
      - payments
  /api/recipients:
    get:
      description: Returns all saved payment recipients for the authenticated client.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List payment recipients
      tags:
      - recipients
    post:
      consumes:
      - application/json
      description: Saves a new payment recipient for the authenticated client.
      parameters:
      - description: Recipient data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePaymentRecipientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a payment recipient
      tags:
      - recipients
  /api/recipients/{id}:
    delete:
      description: Deletes a saved payment recipient. Only the owning client can delete.
      parameters:
      - description: Recipient ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a payment recipient
      tags:
      - recipients
    put:
      consumes:
      - application/json
      description: Updates a saved payment recipient. Only the owning client can update.
      parameters:
      - description: Recipient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Recipient data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePaymentRecipientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a payment recipient
      tags:
      - recipients
  /api/transfers:
    post:
      consumes:
      - application/json
      description: Transfers funds between two accounts belonging to the authenticated
        client. Applies currency conversion when accounts have different currencies.
      parameters:
      - description: Transfer data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTransferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Transfer between own accounts
      tags:
      - transfers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"